Autoupdate of 2 submodules
*trusty/user/base updated to 420fbf3d09357818ea2b04bcd48369951cd8c25b
trusty: prepare_dma: add test for invalid arguments

Bug: 268393649
Test: build.py && trusty-ut-ctrl com.android.libctest
Change-Id: I2e424be649169e0beff0908a58d6f0d18b7417bf

*trusty/kernel updated to 7a83102cf2a0844f9bc0635c1dcd1c27a666d875
trusty: prepare_dma: allow pmem parameter to be omitted.

In some cases the prepare_dma / finish_dma API may be used for cache
ops, but depending on how / where the memory was allocated, the physical
addresses may not be required.  In such a case the `pmem` parameter is
an inconvenience as it has to be allocated to the worst case memory
fragmentation based on the buffer size, and also has a small performance
impact.

Testing shows that omitting the pmem parameter speeds up prepare_dma()
by between 6 and 14%, depending on the size and direction of the
preparation.

Test: build.py
Bug: 268393649
Change-Id: Ic19e7a766933d2ef8595db81ba319d6b45143520