Autoupdate of 2 submodules
*external/trusty/lk updated to a3762ffc519eda735ea23e7d5c166f8dd73bbc6f
dev: interrupt: arm_gic: Use EOIMode=1 to defer irqs in doorbell mode

The current mechanism for deferring Secure IRQ work when taking an FIQ
is brittle in the face of priority masking, where the PMR isn't
necessarily context-switched by TF-A and therefore the priority of
Secure interrupts has to be chosen carefully so as not to be masked by
Linux in the non-secure world

The simplest way to resolve this problem is to move Secure interrupts
into the Secure priority range, however this interacts badly with the
logic to defer IRQ handling from FIQ context when the Secure interrupt
fires while running in the Non-Secure world. In this case, the
higher-priority Secure interrupt is left masked when returning to Linux,
preventing any Non-Secure interrupts from being handled.

So, let's bite the bullet and move to EOIMode=1 on the Secure side. This
splits the priority drop from the deactivation so that the FIQ handler
can ACK+EOI each pending Secure interrupt before injecting the doorbell
interrupt to the Non-Secure world. Later, when the NOP SMC is issued,
these deferred active interrupts can be handled and deactivated without
the need for playing games with group masking and priorities.

Bug: 308278090
Test: on device
Test: trusty-nop thread no longer spinning when pNMIs enabled in Linux
Signed-off-by: Will Deacon <willdeacon@google.com>
Change-Id: Ib9cf3f038dd48da4b0ebace21c1013eca8ab6c66

*trusty/kernel updated to 89c638ffaac6d7f7b86e6163dc59130c05f1436a
trusty: sm: Delineate critical section in idle thread

Introduce a couple of helpers to delineate the critical section of the
low-priority idle thread handling stdcalls.

No functional change.

Bug: 308278090
Test: on device
Signed-off-by: Will Deacon <willdeacon@google.com>
Change-Id: I6408a5cc2167099e90c030995e3f76937079b4fd