run-test: Create runner scripts as part of the build.
The generated zip file includes json file which includes
the commands needed to execute the run-tests as bash script.
The CI just needs to push the data to the device and run the script.
This is just the first step, the scripts are currently unused.

This is implemented by running testrunner.py from the soong build step,
with ART_TEST_RUN_FROM_SOONG set so that the script can adjust behaviour
accordingly (adb is not available; nor is the full Android source tree).

For the time being, this generates arm-specific script,
although it should be easy to make the scrip arch-agnostic.

Test: The commands are exactly identical as during test.py execution.

Change-Id: I8427dd68ffab0815f23034920860f2616755d8f8