Define the initial interface for the PinWeaver Storage Daemon
The storage daemon acts as a tipc client, and serves storage
requests from the PinWeaver trusty app. The API as-is is based
on the current design.

Ser/de of complex data using `tipc` is remarkably difficult.
The `util` module enables this for enums, `Vec`, and `Option`.
I also realized, far too late, that `tipc` and `trusty-rs` are
different libraries, and non-Trusty apps don't have access to
the `tipc` ser/de traits. This is why these duplicating traits
exist in `util`: to use the same interface for client and server.

Ideally, this would be replaced with an AIDL interface,
but tipc doesn't support using rpcbinder to connect a Trusty app
to an Android service.

It might be possible to still use AIDL by utilizing a callback
and having the storage daemon connect to the Trusty app. This
should be investigated before more effort is put into this API.

Bug: 352072286
Test: CQ
Change-Id: I46a76400f587f1d50477f0c37b06fe6ed4b5db2f