Autoupdate of 3 submodules
*art updated to c6c8e76a6995acef2861f1a528610375bcaa3052
Explicitly use implementations where necessary for modules in the same apex

The logic that magically determines whether to use stubs or
implementations for modules in the same apex is being removed
to simplify the Soong logic and make the build system easier to
understand.  Explicitly use libdexfile#impl for modules in the
art apex that need the implementation APIs.

Bug: 372543712
Test: builds
Flag: EXEMPT refactor
Change-Id: I916d54fc558d6030bcb3b0586eb0f11a41a24c42

*build/soong updated to 8ffb90f6da1c939a9f6041013d88b71dab7ab88f
Merge "Don't magically use implementations for modules in the same apex" into main
*packages/modules/StatsD updated to 5f058c717380e6fea1bdb0ead7aca7752e52546a
Merge "Explicitly use implementations where necessary for modules in the same apex" into main