Always adjust last usable block when flashing GPT
The GPT image flashed to a device may be generated for a different disk
size. Thus we need to always adjust the last usable block according the
disk size of the target device. This also allows GPT verification to
catch any partition that will be out of range on this device.

Bug: 380129511
Change-Id: I265099c63d23c833e1ca24b574a4c4192250365d